migration to androidx and added arch dependencies
