+++ app/build.gradle
@@ -11,7 +11,7 @@ android {
         targetSdkVersion COMPLILE_SDK_VERSION
         versionCode 1
         versionName "0.4"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         vectorDrawables.useSupportLibrary = true
     }
     //TODO: fix lint issues.
@@ -67,29 +67,49 @@ android {
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    implementation "com.android.support:appcompat-v7:${SUPPORT_TOOL_VERSION}"
-    implementation "com.android.support:support-v4:${SUPPORT_TOOL_VERSION}"
-    implementation "com.android.support:design:${SUPPORT_TOOL_VERSION}"
-    implementation "com.android.support:customtabs:${SUPPORT_TOOL_VERSION}"
-    implementation "com.android.support:recyclerview-v7:${SUPPORT_TOOL_VERSION}"
-    implementation "com.squareup.retrofit2:retrofit:${RETROFIT_VER}"
-    implementation "com.squareup.retrofit2:converter-gson:${RETROFIT_CONV_VER}"
-    implementation "com.squareup.retrofit2:adapter-rxjava2:${RETROFIT_VER}"
-    implementation "com.squareup.okhttp3:logging-interceptor:${RETROFIT_LOGGER_VER}"
+    implementation "androidx.appcompat:appcompat:${ANDROID_X_VERSION}"
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'com.google.android.material:material:1.0.0'
+    implementation "androidx.browser:browser:${ANDROID_X_BROWSER_VERSION}"
+    implementation "androidx.recyclerview:recyclerview:${ANDROID_X_VERSION}"
+    implementation "com.squareup.retrofit2:retrofit:${RETROFIT_VERSION}"
+    implementation "com.squareup.retrofit2:converter-gson:${RETROFIT_CONV_VERSION}"
+    implementation "com.squareup.retrofit2:adapter-rxjava2:${RETROFIT_VERSION}"
+    implementation "com.squareup.okhttp3:logging-interceptor:${RETROFIT_LOGGER_VERSION}"
     implementation "com.google.dagger:dagger:${DAGGER_VERSION}"
     kapt "com.google.dagger:dagger-compiler:${DAGGER_VERSION}"
-    implementation "io.reactivex.rxjava2:rxjava:${RX_JAVA_VER}"
-    implementation "io.reactivex.rxjava2:rxandroid:${RX_ANDROID_VER}"
-    implementation "com.squareup.picasso:picasso:${PICASSO_VER}"
-    implementation "com.github.florent37:materialimageloading:${PICASSO_MATERIAL}"
-    implementation "com.github.esafirm:RxDownloader:${RX_DOWNLOADER}"
+    implementation "io.reactivex.rxjava2:rxjava:${RX_JAVA_VERSION}"
+    implementation "io.reactivex.rxjava2:rxandroid:${RX_ANDROID_VERSION}"
+    implementation "com.squareup.picasso:picasso:${PICASSO_VERSION}"
+    implementation "com.github.florent37:materialimageloading:${PICASSO_MATERIAL_VERSION}"
+    implementation "com.github.esafirm:RxDownloader:${RX_DOWNLOADER_VERSION}"
+
+    /**
+     * Architecture components
+     */
+    //Lifecycle comp:(ViewModel + LiveData)
+    implementation "androidx.lifecycle:lifecycle-extensions:${ARCH_LIFECYCLE_COMPONENT_VERSION}"
+    implementation "androidx.lifecycle:lifecycle-common-java8:${ARCH_LIFECYCLE_COMPONENT_VERSION}"
+    //room + rxjava2
+    implementation "androidx.room:room-runtime:${ARCH_COMPONENT_ROOM_VERSION}"
+    kapt "androidx.room:room-compiler:${ARCH_COMPONENT_ROOM_VERSION}"
+    implementation "androidx.room:room-rxjava2:${ARCH_COMPONENT_ROOM_VERSION}"
+    //paging + rxjava2
+    implementation "androidx.paging:paging-runtime:${ARCH_COMPONENT_PAGING_VERSION}"
+    implementation "androidx.paging:paging-rxjava2:${ARCH_COMPONENT_PAGING_VERSION}"
+
+
+
 
     
-    debugImplementation "com.readystatesoftware.chuck:library:${CHUCK_VER}"
-    releaseImplementation "com.readystatesoftware.chuck:library-no-op:${CHUCK_VER}"
+    debugImplementation "com.readystatesoftware.chuck:library:${CHUCK_VERSION}"
+    releaseImplementation "com.readystatesoftware.chuck:library-no-op:${CHUCK_VERSION}"
 
     //testing
-    testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.1'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.1'
+    testImplementation "junit:junit:${JUNIT_VERSION}"
+    androidTestImplementation "androidx.test.ext:junit:1.1.1"
+    androidTestImplementation "androidx.test.espresso:espresso-core:${ANDROID_X_CORE_TEST_VERSION}"
+    testImplementation "androidx.room:room-testing:${ARCH_COMPONENT_ROOM_VERSION}"
+    testImplementation "androidx.paging:paging-common:${ARCH_COMPONENT_PAGING_VERSION}"
+    testImplementation 'androidx.arch.core:core-testing:2.1.0'
 }
