+++ app/src/main/java/com/karthik/splash/aboutscreen/AboutScreen.kt
@@ -1,10 +1,10 @@
 package com.karthik.splash.aboutscreen
 
 import android.os.Bundle
-import android.support.design.widget.AppBarLayout
-import android.support.v7.app.AppCompatActivity
-import android.support.v7.widget.DividerItemDecoration
-import android.support.v7.widget.LinearLayoutManager
+import com.google.android.material.appbar.AppBarLayout
+import androidx.appcompat.app.AppCompatActivity
+import androidx.recyclerview.widget.DividerItemDecoration
+import androidx.recyclerview.widget.LinearLayoutManager
 import android.view.View
 import com.karthik.splash.models.library.Library
 import com.karthik.splash.R
