+++ app/src/main/java/com/karthik/splash/homescreen/bottomliketab/BottomLikeTabFragment.kt
@@ -2,10 +2,10 @@ package com.karthik.splash.homescreen.bottomliketab
 
 import android.net.Uri
 import android.os.Bundle
-import android.support.customtabs.CustomTabsIntent
-import android.support.v4.app.Fragment
-import android.support.v4.content.ContextCompat
-import android.support.v7.widget.LinearLayoutManager
+import androidx.browser.customtabs.CustomTabsIntent
+import androidx.fragment.app.Fragment
+import androidx.core.content.ContextCompat
+import androidx.recyclerview.widget.LinearLayoutManager
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -90,6 +90,6 @@ class BottomLikeTabFragment: Fragment(),BottomLikeTabContract.view {
         intentBuilder.setSecondaryToolbarColor(ContextCompat.getColor(context!!,
                 R.color.icons))
         val customTabsIntent = intentBuilder.build()
-        customTabsIntent.launchUrl(activity,uri)
+        customTabsIntent.launchUrl(context!!,uri)
     }
 }
