+++ app/src/main/java/com/karthik/splash/homescreen/bottomtab/BottomTabAdapter.kt
@@ -2,8 +2,8 @@ package com.karthik.splash.homescreen.bottomtab
 
 import android.content.Context
 import android.content.Intent
-import android.support.v4.content.ContextCompat
-import android.support.v7.widget.RecyclerView
+import androidx.core.content.ContextCompat
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -78,7 +78,7 @@ class BottomTabAdapter(private val photolist:ArrayList<Photos>,
         }
     }
 
-    inner class ProgressViewHolder(rootview:View):RecyclerView.ViewHolder(rootview)
+    inner class ProgressViewHolder(rootview:View): RecyclerView.ViewHolder(rootview)
 
     private fun loadImage(position:Int,holder:BottomViewHolder,context: Context){
         Picasso.with(context)
