+++ app/src/main/java/com/karthik/splash/homescreen/bottomtab/BottomTabFragment.kt
@@ -1,8 +1,8 @@
 package com.karthik.splash.homescreen.bottomtab
 
 import android.os.Bundle
-import android.support.v4.app.Fragment
-import android.support.v7.widget.LinearLayoutManager
+import androidx.fragment.app.Fragment
+import androidx.recyclerview.widget.LinearLayoutManager
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -12,7 +12,7 @@ import com.karthik.splash.root.SplashApp
 import kotlinx.android.synthetic.main.fragment_new.*
 import javax.inject.Inject
 
-class BottomTabFragment:Fragment(), BottomTabContract.View, PaginatedView {
+class BottomTabFragment: Fragment(), BottomTabContract.View, PaginatedView {
 
     private var bottomTabComponent: BottomTabComponent?=null
     private lateinit var feedsAdapter:BottomTabAdapter
