+++ app/src/main/java/com/karthik/splash/photodetailscreen/PhotoDetailScreen.kt
@@ -3,10 +3,10 @@ package com.karthik.splash.photodetailscreen
 import android.Manifest
 import android.content.Intent
 import android.os.Bundle
-import android.support.v4.app.ActivityCompat
-import android.support.v4.content.ContextCompat
-import android.support.v4.content.PermissionChecker
-import android.support.v7.app.AppCompatActivity
+import androidx.core.app.ActivityCompat
+import androidx.core.content.ContextCompat
+import androidx.core.content.PermissionChecker
+import androidx.appcompat.app.AppCompatActivity
 import android.view.View
 import android.widget.Toast
 import com.karthik.splash.misc.CircularTransform
