--- app/src/main/java/com/karthik/splash/splashscreen/SplashScreenComponent.kt
@@ -1,9 +1,0 @@
-package com.karthik.splash.splashscreen
-
-import dagger.Subcomponent
-
-@SplashScreenScope
-@Subcomponent(modules = [SplashScreenModule::class])
-interface SplashScreenComponent {
-    fun inject(splashscreen:SplashScreen)
-}
