+++ app/src/main/res/layout/activity_about.xml
@@ -1,4 +1,4 @@
-<android.support.design.widget.CoordinatorLayout
+<androidx.coordinatorlayout.widget.CoordinatorLayout
         xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
@@ -6,12 +6,12 @@
         android:background="@color/tab_selected_color"
         android:orientation="vertical">
 
-    <android.support.design.widget.AppBarLayout
+    <com.google.android.material.appbar.AppBarLayout
             android:id="@+id/appbarlayout"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar">
-        <android.support.design.widget.CollapsingToolbarLayout
+        <com.google.android.material.appbar.CollapsingToolbarLayout
                 android:id="@+id/collapsingtoolbar"
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
@@ -19,7 +19,7 @@
                 app:layout_scrollFlags="scroll|exitUntilCollapsed|snap"
                 app:title="@string/about">
 
-            <android.support.v7.widget.Toolbar
+            <androidx.appcompat.widget.Toolbar
                     android:id="@+id/toolbar"
                     android:layout_width="match_parent"
                     android:layout_height="?attr/actionBarSize"
@@ -83,9 +83,9 @@
                         android:layout_height="0.4dp"
                         android:background="@color/colorPrimary"/>
             </LinearLayout>
-        </android.support.design.widget.CollapsingToolbarLayout>
-    </android.support.design.widget.AppBarLayout>
-    <android.support.v7.widget.RecyclerView
+        </com.google.android.material.appbar.CollapsingToolbarLayout>
+    </com.google.android.material.appbar.AppBarLayout>
+    <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/librarieslist"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
@@ -94,4 +94,4 @@
             app:layout_behavior="@string/appbar_scrolling_view_behavior"
             android:layout_marginBottom="@dimen/lib_row_height"/>
 
-</android.support.design.widget.CoordinatorLayout>
+</androidx.coordinatorlayout.widget.CoordinatorLayout>
