+++ app/src/main/res/layout/activity_home.xml
@@ -17,7 +17,7 @@
             android:layout_weight="1"
             android:layout_gravity="center"/>
 
-    <android.support.design.widget.BottomNavigationView
+    <com.google.android.material.bottomnavigation.BottomNavigationView
             android:id="@+id/navigation"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
