+++ app/src/main/res/layout/fragment_new.xml
@@ -5,7 +5,7 @@
              android:orientation="vertical"
              android:layout_width="match_parent"
              android:layout_height="match_parent">
-    <android.support.v7.widget.RecyclerView
+    <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/feedslist"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
