+++ build.gradle
@@ -21,18 +21,24 @@ allprojects {
 }
 
 ext{
-    SUPPORT_TOOL_VERSION='28.0.0'
-    BUTTERKNIFE_VERSION = '8.8.0'
+    JUNIT_VERSION='4.12'
+    EXPRESSO_CORE_VERSION='3.1.0'
     DAGGER_VERSION = '2.10'
     COMPLILE_SDK_VERSION=29
     MIN_SDK_VERSION=21
-    RETROFIT_VER='2.3.0'
-    RETROFIT_CONV_VER='2.2.0'
-    RETROFIT_LOGGER_VER='3.4.1'
-    PICASSO_VER='2.5.2'
-    RX_JAVA_VER='2.1.6'
-    RX_ANDROID_VER='2.0.1'
-    CHUCK_VER='1.1.0'
-    PICASSO_MATERIAL='1.0.2'
-    RX_DOWNLOADER='1.0.1'
+    RETROFIT_VERSION='2.3.0'
+    RETROFIT_CONV_VERSION='2.2.0'
+    RETROFIT_LOGGER_VERSION='3.4.1'
+    PICASSO_VERSION='2.5.2'
+    RX_JAVA_VERSION='2.1.6'
+    RX_ANDROID_VERSION='2.0.1'
+    CHUCK_VERSION='1.1.0'
+    PICASSO_MATERIAL_VERSION='1.0.2'
+    RX_DOWNLOADER_VERSION='1.0.1'
+    ARCH_LIFECYCLE_COMPONENT_VERSION='2.2.0-rc03'
+    ARCH_COMPONENT_ROOM_VERSION='2.2.3'
+    ARCH_COMPONENT_PAGING_VERSION='2.1.1'
+    ANDROID_X_VERSION='1.1.0'
+    ANDROID_X_BROWSER_VERSION='1.2.0'
+    ANDROID_X_CORE_TEST_VERSION='3.1.0'
 }
