+++ gradle.properties
@@ -9,6 +9,8 @@
 
 # Specifies the JVM arguments used for the daemon process.
 # The setting is particularly useful for tweaking memory settings.
+android.enableJetifier=true
+android.useAndroidX=true
 org.gradle.daemon=true
 org.gradle.jvmargs=-Xmx1536m
 org.gradle.parallel=true
